+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - bug fixes
   - regression preventing glass effect from using the widget's color
+- widgets
+  - text: set line-height relative to font size
 
 ## 0.26.1