+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - bug fixes
   - regression preventing glass effect from using the widget's color
+  - missing last key in keyboard widget
 - widgets
   - text: set line-height relative to font size